backend/requirements.txt
asgiref==3.9.1
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.8.3
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
constantly==23.10.4
cryptography==45.0.7
daphne==4.2.1
defusedxml==0.7.1
Django==5.2.5
django-cors-headers==4.7.0
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-nested-routers==0.94.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.2
oauthlib==3.3.1
pillow==11.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.3.0
python3-openid==3.2.0
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==80.9.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
zope.interface==8.0.1
